+++ OnlineCinema.BLL/OnlineCinema.BLL.csproj
@@ -1,0 +1,18 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+    <PropertyGroup>
+        <TargetFramework>net6.0</TargetFramework>
+        <ImplicitUsings>enable</ImplicitUsings>
+        <Nullable>enable</Nullable>
+    </PropertyGroup>
+
+    <ItemGroup>
+      <PackageReference Include="Microsoft.EntityFrameworkCore.Abstractions" Version="6.0.5" />
+    </ItemGroup>
+
+    <ItemGroup>
+      <ProjectReference Include="..\OnlineCinema.DAL\OnlineCinema.DAL.csproj" />
+      <ProjectReference Include="..\OnlineCinema.Domain\OnlineCinema.Domain.csproj" />
+    </ItemGroup>
+
+</Project>
